UI and Gameplay elements
The gameplay elements are set up but no code has been written yet. The next step for the template is to make set the ui activeness in the awake function of a scene controller script. After that ui buttons should be made to work. The timer then can be implemented but only start after first input and reset after a restart. I will be back soon to write code for the ui and set up the gameplay code.